+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 
 - Fix double quoted email address validation.
   [chris-adam]
-
+- Used also `security.is_develop_environment` in `path.is_test_url`.
+  [sgeulette]
 
 1.3.14 (2026-04-14)
 -------------------